Fix compiler warnings about narrowing, name hiding, unused variables or parameters. Silence some msvc warnings.
(cherry picked from commit 14a08d521a5155f28a75f56064360ea6e91a37cd)

+++ include/boost/geometry/formulas/area_formulas.hpp
@@ -567,7 +567,7 @@ public:
             state.m_crosses_prime_meridian++;
         }
 
-        return state.m_crosses_prime_meridian;
+        return static_cast<int>(state.m_crosses_prime_meridian);
     }
 
 };
